params for miniPairs added

+++ njst/src/njst.h
@@ -3,6 +3,6 @@
 #ifndef NJST_H
 #define NJST_H
 
-void inferSpeciesTreeFromGeneTrees(struct node **speciesTree, const char *filename, int mini, int ustar, int norm, int branchLength, int weight, int square, int tag, int root, int miniPairs);
+void inferSpeciesTreeFromGeneTrees(struct node **speciesTree, const char *filename, int mini, int ustar, int norm, int branchLength, int weight, int square, int tag, int root, double miniPairs);
 
 #endif
